Move advanced cost model to separate crate (and re-test old base cost model)
