+++ optd-datafusion-repr-adv-cost/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "optd-datafusion-repr-adv-cost"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+optd-datafusion-repr = { path = "../optd-datafusion-repr" }
+optd-core = { path = "../optd-core" }
