+++ optd-datafusion-repr-adv-cost/src/lib.rs
@@ -1,0 +1,5 @@
+mod adv_cost;
+
+pub trait WithRuntimeStatistics {
+    fn get_runtime_statistics(&self) -> RuntimeAdaptionStorage;
+}
